[+]main page metadata
