+++ src/app/actions/authCheckerOnLayout.tsx
@@ -1,0 +1,16 @@
+'use client'
+import { useAuthStore } from "~/store/authStore";
+import { useEffect } from "react";
+
+export const AuthCheckerOnLayout : React.FC = () => { 
+    const checkAuth = useAuthStore((state) => state.checkAuth);
+    useEffect(()  => {
+        const fetchData = async () => {
+        await checkAuth(); // Загружаем пользователя при старте
+        }
+        void fetchData()
+    }, [checkAuth]);
+    return (
+        <></>
+    )
+    }
